assemble-app: Extract AppHelper, return app object
Makes assemble-app-test.js easier to read in preparation for testing
more API methods.
