+++ lib/index.js
@@ -28,4 +28,5 @@ function assembleApp(redirectDb, app) {
   app.put('/', function(req, res) {
   })
   */
+  return app
 }
